requirements.txt
ultralytics==8.0.204
pygame==2.5.2
numpy==1.26.4
opencv-python==4.9.0.80
RPi.GPIO==0.7.1
